Add ksp prebuilts to ksp directory.
We do not have the dependencies necessary to compile ksp
from source. This is particularly apparent when viewing
kotlin-analysis-api/build.gradle.kts, which has a great number
of non-ksp dependencies.

The followup cl to this will remove all the existing source.

Attestations from the kokoro build can be found at
/placer/prod/home/kokoro-dedicated/build_artifacts/prod/kotlin-symbol-processing/gcp_ubuntu/release/20250910-115439/4cee4621-759e-4ea3-b9d2-5a7ac18994e9/jars/github/ksp/

Bug: 442809933
Test: m SystemUI with ksp enabled.
Change-Id: I270091e95f9c7cc93bd36696313c0a535be0f4bc